add prelude module and tests
